ci: clang-format pass + uint8_t->uint32_t loop var in test_rx_poeg
Cleanup of two CI failures from the cross-check commit (cb2999898):
- clang-tidy bugprone-too-small-loop-variable: test_rx_poeg.c:393 used
  uint8_t i for a loop bounded by k_poeg_motor_count (untyped enum =
  int). Widened to uint32_t to match the other loops in the file.
- clang-format-18 wanted minor reflow on adc.c, rx_mpc.c, rx_mtu.c,
  rx_poeg.c after the cross-check edits. Pure formatting.

55/55 host tests still pass.

+++ star-rx72n-firmware/tests/test_rx_poeg.c
@@ -390,7 +390,7 @@ static void test_poeg_isr_handler(uint8_t motor_index, uint16_t unused_vector)
   const uint32_t grpbl2 = (1UL << s_poeg_grpbl2_bits_test[motor_index]);
 
   bool faulted = false;
-  for (uint8_t i = 0; i < k_poeg_motor_count; i++) {
+  for (uint32_t i = 0; i < k_poeg_motor_count; i++) {
     if ((grpbl2 & (1UL << s_poeg_grpbl2_bits_test[i])) == 0U) {
       continue;
     }
